added new lines

+++ simple_calculator.py
@@ -22,8 +22,8 @@ print("4. Divide")
 choice = input("Enter choice (1/2/3/4): ")
 
 try:
-    num1 = float(input("Enter first number: "))
-    num2 = float(input("Enter second number: "))
+    num1 = int(input("Enter first number: "))
+    num2 = int(input("Enter second number: "))
 
     if choice == '1':
         print("Result:", add(num1, num2))
